Fixes wrong paths
For some reasons, that regex broke the SCSS includes. It was marked as a regression fix on 1a17a9b7ebb1e3b0084b8762a2998b85765faa26 but... i'm not sure what it fixes there :)

The problem I suspect is that it is augmenting the path like this:

```
[
"'pages/article'", 
"'pages/article'pages/article",  <<<<<<<<<<<<<<<<
...]
```

And later on the parser has trouble with that.

PS: while the opening existing files didn't worked, the creation of new files did.